% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grib-to-netcdf.R
\name{grib.convert}
\alias{grib.convert}
\title{GRIB to NETCDF conversion}
\usage{
grib.convert(fnm)
}
\arguments{
\item{fnm}{Filename to be converted (including full path)}
}
\description{
Creates a new file containing the specified data, converted to NETCDF format
}
\details{
Downloaded data must be converted to .nc format before importing with the other functions. This is a wrapper function to call 'grib_to_netcdf' from 'eccodes' package without having to go to the command line
}

