% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{land.sea.mask}
\alias{land.sea.mask}
\title{Import land-sea mask}
\usage{
land.sea.mask(fnm = "./01-raw-data/invariant/ERA-int-lsm.nc", N = 60,
  W = -6, S = 50, E = 2)
}
\arguments{
\item{fnm}{Filename containing .nc land-sea mask. Defaults to last storage location.}

\item{NWSE}{Four separate parameters indicating N,W,S,E boundaries of region of interest (in degrees). Default selects a small region centred on the UK.}
}
\value{
Labelled array of 0 (= sea) or 1 (= land)
}
\description{
Import the invariant land-sea mask used by ERA-interim from a netcdf download
}

