% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{lat}
\alias{lat}
\title{Get latitudes}
\usage{
lat(arr)
}
\arguments{
\item{arr}{Array from which latitude is to be extracted}
}
\value{
Vector of numeric latitude values
}
\description{
Plotting function to concisely extract latitude range from an array with named dimensions
}

