% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{lon}
\alias{lon}
\title{Get longitudes}
\usage{
lon(arr)
}
\arguments{
\item{arr}{Array from which longitude is to be extracted}
}
\value{
Vector of longitude values
}
\description{
Plotting function to concisely extract longitude range from an array with named dimensions
}

