% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-netcdf.R
\name{nc.array}
\alias{nc.array}
\title{Load netcdf file into array}
\usage{
nc.array(fnm, lt.as.days = F)
}
\arguments{
\item{fnm}{Filename of netcdf object to import (including path)}

\item{lt.as.days}{Boolean: return forecast step/leadtime in hours (F) or days (T)? Default is hours)}
}
\value{
Array of same dimensions as source data, with all dimensions labelled according to the netcdf specifications.
}
\description{
Import the contents of a .nc file into an array of the specified dimensions
}

