% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-netcdf.R
\name{nc.inspect}
\alias{nc.inspect}
\title{Inspect netcdf file}
\usage{
nc.inspect(fnm, detail = T)
}
\arguments{
\item{fnm}{Filename of netcdf object to inspect (including path)}

\item{detail}{Boolean: report only names of dimensions (F) or their values (T)? Default is to show all values.}
}
\description{
Inspect a .nc file and print a list of the variables and dimensions.
}

