% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masks.R
\name{ns.mask}
\alias{ns.mask}
\title{Create north-south mask}
\usage{
ns.mask()
}
\value{
List of labelled land-sea masks, with sea values marked as NA. Each mask can therefore be multiplied by a temperature field of the same dimensions to obtain only the values of interest.
}
\description{
Define the regions considered as 'north UK' and 'south UK' in the study
}
\details{
The 'northern' region consists of all land regions above 54.5 degrees north, with the 'southern' region containing all land regions in the UK below this latitude.
}

